+++ lib/Base.php
@@ -356,6 +356,8 @@ class Base implements LoggerAwareInterface
             $status_bin = '';
             $status = '';
             // Get the close status.
+            $status_bin = '';
+            $status = '';
             if ($payload_length > 0) {
                 $status_bin = $payload[0] . $payload[1];
                 $status = current(unpack('n', $payload));
@@ -412,12 +414,9 @@ class Base implements LoggerAwareInterface
         $length = strlen($data);
         $written = fwrite($this->socket, $data);
         if ($written === false) {
-            fclose($this->socket);
             $this->throwException("Failed to write {$length} bytes.");
         }
-
         if ($written < strlen($data)) {
-            fclose($this->socket);
             $this->throwException("Could only write {$written} out of {$length} bytes.");
         }
         $this->logger->debug("Wrote {$written} of {$length} bytes.");
@@ -430,11 +429,9 @@ class Base implements LoggerAwareInterface
             $buffer = fread($this->socket, $length - strlen($data));
             if ($buffer === false) {
                 $read = strlen($data);
-                fclose($this->socket);
                 $this->throwException("Broken frame, read {$read} of stated {$length} bytes.");
             }
             if ($buffer === '') {
-                fclose($this->socket);
                 $this->throwException("Empty read; connection dead?");
             }
             $data .= $buffer;
@@ -448,6 +445,7 @@ class Base implements LoggerAwareInterface
     {
         $meta = $this->isConnected() ? stream_get_meta_data($this->socket) : [];
         $json_meta = json_encode($meta);
+        fclose($this->socket);
         if (!empty($meta['timed_out'])) {
             $this->logger->error($message, $meta);
             throw new TimeoutException($message, ConnectionException::TIMED_OUT, $meta);