+++ docs/Changelog.md
@@ -9,6 +9,7 @@
 ### `1.5.0`
 
  * Convenience send methods; text(), binary(), ping(), pong() (@sirn-se)
+ * Optional Message instance as receive() method return (@sirn-se)
  * Opcode filter for receive() method (@sirn-se)
  * Fix for unordered framgemented messages (@sirn-se)
  * Various code re-write (@sirn-se)